Add files via upload
The increasing number of cancellations calls for a Machine Learning based solution that can help in predicting which booking is likely to be canceled. Have to analyze the data provided to find which factors have a high influence on booking cancellations, build a predictive model that can predict which booking is going to be canceled in advance, and help in formulating profitable policies for cancellations and refunds.